[Feature] : play sound on keypress
First check if the pressed key is acceptable.
This is done by comparing the keyCode with an array of acceptable keyCodes.
Then find the audio using querySelectorAll.
Find the element with audio tag that contains data-key attribute with value of the keyCode.
Then play it and the corresponding sound will play.