[Refactor] : change querySelector condition from div to .key
Instead of querying for div elements with data-key attributes,
change so that it queries for elements with className key and data-key attributes.
Thought in this file there are no problem using div elements,
it seems more accurate to use .key, because if there were more div attributes
without the className key, then an error might occur.