[Refactor] : change method of checking the pressed key
Instead of creating an array of acceptable keyCodes and
comparing it with the event keyCode,
find the audio element with the data-set attribute value to the keyCode.
If an element is not found, then the pressed key is not acceptable.
Returning when not found will prevent the rest of the code from running.