[Refactor] : change method of removing playing class
SetTimeout might not match well with the transition time.
So instead, add an event on transitionend.
First, get all the element with className key using querySelectorAll.
Iterate the element in the array and add an transisionend event listener.
There are multiple transitions ending: border, size, transform, etc.
Only continue when the property Name of the event is transform.
And then remove the class playing.