[Feature] : update clock hand position
Get the hours, minutes, seconds of the current time.
Get the clock hands using querySelector.
Transform the origin of the hand to 100%.
This is because the transfor origin is the center of the element by default.
Totate each hand correspondingly.
Each hour needs to rotate for 30 degrees, while each minutes for 6 degrees.
Add 90 degrees since the hand starts at 270 degrees.
Repeate the process every second using setInterval.