[Feature] : calculate the total years lived by the inventors using reduce
Create a reducer function that accumulates the total years lived by the inventors.
Get the year lived by subtracting the birth year to the passed year.
Then add it to the accumulator so that the year lived gets accumulated.
Place an initial value ton the reduce function's parameter.
If not, the first element of the array will be passed as the accumulator, which is an object,
hence resulting in a string with all the year lived concatenated.