[Feature] : sort people alphabetically by last name
Create a compare function that gets the last name from the fullname and compares it.
Split the full name by ', ' and get the second element of the result; the last name.
If the last name of a is alphabetically bigger than the last name of b, return 1, and vice versa.
If they are equal, return 0.