[Feature] : list of boulevards in Paris that contain 'de'
Get all the a tags that contains the name of boulevards in paris using querySelectorAll.
The result returns a nodeList. In order to use the method map, it needs to be an array.
Therefore, using destructuring, convert the nodeList into an array.
Using map(), only retrieve the title of the a tag element.
Using reduce, if a boulevard contains 'de', then add it  to the accumulating array.
Return the array in order to update it.