[Refactor] : change how items are aligned
Instead of using the space-evely property, use flex auto and display flex.
Space-evenly places elements evenly, having the element's size as it is.
By using flex, the element actually take the space that is available.
Therefore, later when using translateY(-100%), it will be able to
actually place it outside the window.