[Refactor] : add transitionend event listener
Before changing, the flex-grow and translate transform occurs at the same time.
In order to make the translate transform occur after the flex-grow transform,
add an event listener of type transitionend to each pannel.
However, when a transition occurs, multiple transformations occur at the same time.
Therefore , by using e.propertyName, only toggle the className open-active
when the flex-grow transition has occurred.

Instead of checking if the propertyName is flex-grow, check whether the propertyName
contains the word flex. This is because in safari, the flex transformation's name
is flex, while in other browsers, its flex-grow.