[Refactor] : change event listener type
The event type input only listens to characters, numbers, and special keys.
On the other hand, the keyup type also listens to control keys like enter, ctrl, backspace, delete, etc.
Because we need to detect the backspace also, the keyup type should be used.
Also, the reason why it should be keyup instead of keydown is because
keydown event is fired before the new character is added to the value.
Since we need to use the value inserted by the pressed key, keyup should be used.