[Feature] : highlight the search keyword in the results
Everytime the search keyword is inserted, it must ccreate
a new list with the new highlight on it.
Therefore, instead of creating the list elements before hand,
create it everytime the key is pressed.
Check if the city of state contains the keyword.
If it does, wrap the match with the highlighting tag.
Add this to the suggestions element.